+++ CHANGELOG.md
@@ -13,6 +13,10 @@ As part of ongoing work to improve our developer APIs, we've redesigned and docu
 - CA Category now has a brand_id property, and the "colour" key has been renamed to "color"
 - Site addresses now have latitude and longitude keys
 - The CA Embed Configuration model no longer has id/mtime/muser/ctime/cuser/resource_module properties, and the "type" property has been renamed to "format"
+- new sequenceId, identifier and meetingUrl properties have been added to Event
+- the image properties on Event is now an object with multiple size options
+- the _original property on Event has been removed
+- the siteIds property for Event now contains an array of site IDs for All Sites models rather than being null, and a new allSites boolean property has been added
 
 ## [4.3.7] - 2024-09-02